% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fn.csv2sql}
\alias{fn.csv2sql}
\title{Import CSV file to SQLite database}
\usage{
fn.csv2sql(file, sep, encoding = "UTF-8", database, table, index)
}
\arguments{
\item{file}{CSV: Filepath to CSV file with source data}

\item{sep}{CSV: Field seperator used in CSV file}

\item{encoding}{CSV: Encoding used in CSV file (default: UTF-8)}

\item{database}{SQL: SQLite database connection name}

\item{table}{SQL: Target SQLite table name}

\item{index}{SQL: Field name(s) to create an index for in the SQLite database}
}
\description{
Import CSV file to SQLite database
}
\examples{
fn.csv2sql("raw_data/mycsv.csv", ";", "UTF-8", sqllite_connection, "mysqltable", "id")
fn.csv2sql("raw_data/mycsv.csv", ";", "UTF-8", sqllite_connection, "mysqltable", c("id1","id2"))

}
