% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fn.df2sql}
\alias{fn.df2sql}
\title{Import data frame to SQLite database}
\usage{
fn.df2sql(dataframe, database, table, index)
}
\arguments{
\item{dataframe}{Data frame: Source data stored in a data frame}

\item{database}{SQL: SQLite database connection name}

\item{table}{SQL: Target SQLite table name}

\item{index}{SQL: Field name(s) to create an index for}
}
\description{
Import data frame to SQLite database
}
\examples{
fn.df2sql(mydf, sqllite_connection, "mysqltable", "id")
fn.df2sql(mydf, sqllite_connection, "mysqltable", c("id1","id2"))

}
