% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fn.mssql}
\alias{fn.mssql}
\title{Get data from Microsoft SQL Server database}
\usage{
fn.mssql(server, database, query)
}
\arguments{
\item{server}{Microsoft SQL Server (host name or IP address)}

\item{database}{SQL database name}

\item{query}{SQL query}
}
\value{
Data set collected by SQL query
}
\description{
Establishes a connection to a Microsoft SQL Server database, executes a query
and returns the data collected by the query.
}
\examples{
fn.mssql("hostname", "database", "select * from table")
fn.mssql("sqlserver2019", "empire", "select * from customer")

}
