% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fn.packages}
\alias{fn.packages}
\title{Package loader}
\usage{
fn.packages(pkg)
}
\arguments{
\item{pkg}{A vector with one or more package names.}
}
\description{
This function loads R packages into your work environment.
When a package is not yet installed on your computer, it will be downloaded and installed.
}
\examples{
fn.packages(c("tidyverse"))
fn.packages(c("tidyverse", "knitr"))

}
