% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fn.sqlstructure}
\alias{fn.sqlstructure}
\title{Get structure of SQLite table}
\usage{
fn.sqlstructure(database, table)
}
\arguments{
\item{database}{SQLite database connection name}

\item{table}{SQLite table name}
}
\value{
SQLite table structure with field name(s) and data type returned as data frame
}
\description{
Get structure of SQLite table
}
\examples{
fn.sqlstructure(sqllite_connection, "mysqltable")

}
